rename platform_settings
